Update to divide classes in PagingManager to separate files
